refactor: :recycle: Split sign in (#366)
* refactor: :recycle: rename api.py to sign_in.py

* refactor: remove api code from sign_in.py

* refactor: :recycle: remove sign_in code from api.py

* fix: :bug: Fix sign_in module name conflict with sign_in function

* moved sign_in module to sign_in_mod

* refactor: :recycle: rename sign_in_mod to signIn.py

* refactor: :recycle: change sign_in_mod references to signIn
