+++ mintapi/__init__.py
@@ -1,6 +1,7 @@
 import logging
 
 from mintapi.api import *
+from mintapi.signIn import *
 
 
 logging.getLogger("mintapi").setLevel(logging.INFO)
