+++ tests/test_endpoints_sign_in.py
@@ -4,6 +4,7 @@ import pytest
 
 import mintapi
 import mintapi.api
+import mintapi.signIn
 
 from tests.test_driver import category_example
 
@@ -32,7 +33,7 @@ pytestmark = pytest.mark.skipif(
 @pytest.fixture
 def get_mint_driver() -> mintapi.Mint:
     mint = mintapi.Mint()
-    mint.driver = mintapi.api._create_web_driver_at_mint_com(
+    mint.driver = mintapi.signIn._create_web_driver_at_mint_com(
         HEADLESS,
         SESSION_PATH,
         USE_CHROMEDRIVER_ON_PATH,
@@ -43,7 +44,7 @@ def get_mint_driver() -> mintapi.Mint:
 
 @pytest.fixture
 def do_sign_in(get_mint_driver: mintapi.Mint) -> mintapi.Mint:
-    mintapi.api.sign_in(
+    mintapi.sign_in(
         USERNAME,
         PASSWORD,
         get_mint_driver.driver,
@@ -55,7 +56,7 @@ def do_sign_in(get_mint_driver: mintapi.Mint) -> mintapi.Mint:
 
 
 def test_sign_in(get_mint_driver: mintapi.Mint):
-    mintapi.api.sign_in(
+    mintapi.sign_in(
         USERNAME,
         PASSWORD,
         get_mint_driver.driver,
